return int charactersize
because that's what the sfml function does

+++ SFLCARS-interface/String.cpp
@@ -54,7 +54,7 @@ void String::setCharacterSize(float newSize)
 	text.setCharacterSize(newSize);
 }
 
-float String::getCharacterSize() const
+int String::getCharacterSize() const
 {
 	return text.getCharacterSize();
 }
